3619: Rename EntityAttributeModel files
